Move tracetest to oteltest
