+++ oteltest/doc.go
@@ -12,17 +12,5 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package tracetest
-
-import (
-	"time"
-
-	"go.opentelemetry.io/otel/label"
-)
-
-// Event encapsulates the properties of calls to AddEvent or AddEventWithTimestamp.
-type Event struct {
-	Timestamp  time.Time
-	Name       string
-	Attributes map[label.Key]label.Value
-}
+// Package oteltest provides testing utilities for the otel package.
+package oteltest // import "go.opentelemetry.io/otel/oteltest"
