+++ oteltest/mock_tracer_test.go
@@ -12,15 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package tracetest_test
+package oteltest_test
 
 import (
 	"os"
 	"testing"
 	"unsafe"
 
-	"go.opentelemetry.io/otel/api/trace/tracetest"
 	ottest "go.opentelemetry.io/otel/internal/testing"
+	"go.opentelemetry.io/otel/oteltest"
 )
 
 // Ensure struct alignment prior to running tests.
@@ -28,7 +28,7 @@ func TestMain(m *testing.M) {
 	fields := []ottest.FieldOffset{
 		{
 			Name:   "MockTracer.StartSpanID",
-			Offset: unsafe.Offsetof(tracetest.MockTracer{}.StartSpanID),
+			Offset: unsafe.Offsetof(oteltest.MockTracer{}.StartSpanID),
 		},
 	}
 	if !ottest.Aligned8Byte(fields, os.Stderr) {
