+++ oteltest/span_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package tracetest_test
+package oteltest_test
 
 import (
 	"context"
@@ -23,16 +23,16 @@ import (
 	"time"
 
 	"go.opentelemetry.io/otel/api/trace"
-	"go.opentelemetry.io/otel/api/trace/tracetest"
 	"go.opentelemetry.io/otel/codes"
 	"go.opentelemetry.io/otel/internal/matchers"
 	ottest "go.opentelemetry.io/otel/internal/testing"
 	"go.opentelemetry.io/otel/label"
+	"go.opentelemetry.io/otel/oteltest"
 )
 
 func TestSpan(t *testing.T) {
 	t.Run("#Tracer", func(t *testing.T) {
-		tp := tracetest.NewTracerProvider()
+		tp := oteltest.NewTracerProvider()
 		t.Run("returns the tracer used to start the span", func(t *testing.T) {
 			t.Parallel()
 
@@ -46,7 +46,7 @@ func TestSpan(t *testing.T) {
 	})
 
 	t.Run("#End", func(t *testing.T) {
-		tp := tracetest.NewTracerProvider()
+		tp := oteltest.NewTracerProvider()
 		t.Run("ends the span", func(t *testing.T) {
 			t.Parallel()
 
@@ -55,7 +55,7 @@ func TestSpan(t *testing.T) {
 			tracer := tp.Tracer(t.Name())
 			_, span := tracer.Start(context.Background(), "test")
 
-			subject, ok := span.(*tracetest.Span)
+			subject, ok := span.(*oteltest.Span)
 			e.Expect(ok).ToBeTrue()
 
 			e.Expect(subject.Ended()).ToBeFalse()
@@ -86,7 +86,7 @@ func TestSpan(t *testing.T) {
 			tracer := tp.Tracer(t.Name())
 			_, span := tracer.Start(context.Background(), "test")
 
-			subject, ok := span.(*tracetest.Span)
+			subject, ok := span.(*oteltest.Span)
 			e.Expect(ok).ToBeTrue()
 
 			subject.End()
@@ -109,7 +109,7 @@ func TestSpan(t *testing.T) {
 			tracer := tp.Tracer(t.Name())
 			_, span := tracer.Start(context.Background(), "test")
 
-			subject, ok := span.(*tracetest.Span)
+			subject, ok := span.(*oteltest.Span)
 			e.Expect(ok).ToBeTrue()
 
 			expectedEndTime := time.Now().AddDate(5, 0, 0)
@@ -125,7 +125,7 @@ func TestSpan(t *testing.T) {
 	})
 
 	t.Run("#RecordError", func(t *testing.T) {
-		tp := tracetest.NewTracerProvider()
+		tp := oteltest.NewTracerProvider()
 		t.Run("records an error", func(t *testing.T) {
 			t.Parallel()
 
@@ -152,13 +152,13 @@ func TestSpan(t *testing.T) {
 				tracer := tp.Tracer(t.Name())
 				ctx, span := tracer.Start(context.Background(), "test")
 
-				subject, ok := span.(*tracetest.Span)
+				subject, ok := span.(*oteltest.Span)
 				e.Expect(ok).ToBeTrue()
 
 				testTime := time.Now()
 				subject.RecordError(ctx, s.err, trace.WithErrorTime(testTime))
 
-				expectedEvents := []tracetest.Event{{
+				expectedEvents := []oteltest.Event{{
 					Timestamp: testTime,
 					Name:      "error",
 					Attributes: map[label.Key]label.Value{
@@ -180,7 +180,7 @@ func TestSpan(t *testing.T) {
 			tracer := tp.Tracer(t.Name())
 			ctx, span := tracer.Start(context.Background(), "test")
 
-			subject, ok := span.(*tracetest.Span)
+			subject, ok := span.(*oteltest.Span)
 			e.Expect(ok).ToBeTrue()
 
 			errMsg := "test error message"
@@ -189,7 +189,7 @@ func TestSpan(t *testing.T) {
 			expStatusCode := codes.Error
 			subject.RecordError(ctx, testErr, trace.WithErrorTime(testTime), trace.WithErrorStatus(expStatusCode))
 
-			expectedEvents := []tracetest.Event{{
+			expectedEvents := []oteltest.Event{{
 				Timestamp: testTime,
 				Name:      "error",
 				Attributes: map[label.Key]label.Value{
@@ -209,7 +209,7 @@ func TestSpan(t *testing.T) {
 			tracer := tp.Tracer(t.Name())
 			ctx, span := tracer.Start(context.Background(), "test")
 
-			subject, ok := span.(*tracetest.Span)
+			subject, ok := span.(*oteltest.Span)
 			e.Expect(ok).ToBeTrue()
 
 			subject.End()
@@ -226,7 +226,7 @@ func TestSpan(t *testing.T) {
 			tracer := tp.Tracer(t.Name())
 			ctx, span := tracer.Start(context.Background(), "test")
 
-			subject, ok := span.(*tracetest.Span)
+			subject, ok := span.(*oteltest.Span)
 			e.Expect(ok).ToBeTrue()
 
 			subject.RecordError(ctx, nil)
@@ -236,7 +236,7 @@ func TestSpan(t *testing.T) {
 	})
 
 	t.Run("#IsRecording", func(t *testing.T) {
-		tp := tracetest.NewTracerProvider()
+		tp := oteltest.NewTracerProvider()
 		t.Run("returns true", func(t *testing.T) {
 			t.Parallel()
 
@@ -250,7 +250,7 @@ func TestSpan(t *testing.T) {
 	})
 
 	t.Run("#SpanContext", func(t *testing.T) {
-		tp := tracetest.NewTracerProvider()
+		tp := oteltest.NewTracerProvider()
 		t.Run("returns a valid SpanContext", func(t *testing.T) {
 			t.Parallel()
 
@@ -275,7 +275,7 @@ func TestSpan(t *testing.T) {
 	})
 
 	t.Run("#Name", func(t *testing.T) {
-		tp := tracetest.NewTracerProvider()
+		tp := oteltest.NewTracerProvider()
 		t.Run("returns the most recently set name on the span", func(t *testing.T) {
 			t.Parallel()
 
@@ -285,7 +285,7 @@ func TestSpan(t *testing.T) {
 			originalName := "test"
 			_, span := tracer.Start(context.Background(), originalName)
 
-			subject, ok := span.(*tracetest.Span)
+			subject, ok := span.(*oteltest.Span)
 			e.Expect(ok).ToBeTrue()
 
 			e.Expect(subject.Name()).ToEqual(originalName)
@@ -308,7 +308,7 @@ func TestSpan(t *testing.T) {
 			originalName := "test"
 			_, span := tracer.Start(context.Background(), originalName)
 
-			subject, ok := span.(*tracetest.Span)
+			subject, ok := span.(*oteltest.Span)
 			e.Expect(ok).ToBeTrue()
 
 			subject.End()
@@ -319,7 +319,7 @@ func TestSpan(t *testing.T) {
 	})
 
 	t.Run("#Attributes", func(t *testing.T) {
-		tp := tracetest.NewTracerProvider()
+		tp := oteltest.NewTracerProvider()
 		t.Run("returns an empty map by default", func(t *testing.T) {
 			t.Parallel()
 
@@ -328,7 +328,7 @@ func TestSpan(t *testing.T) {
 			tracer := tp.Tracer(t.Name())
 			_, span := tracer.Start(context.Background(), "test")
 
-			subject, ok := span.(*tracetest.Span)
+			subject, ok := span.(*oteltest.Span)
 			e.Expect(ok).ToBeTrue()
 
 			e.Expect(subject.Attributes()).ToEqual(map[label.Key]label.Value{})
@@ -342,7 +342,7 @@ func TestSpan(t *testing.T) {
 			tracer := tp.Tracer(t.Name())
 			_, span := tracer.Start(context.Background(), "test")
 
-			subject, ok := span.(*tracetest.Span)
+			subject, ok := span.(*oteltest.Span)
 			e.Expect(ok).ToBeTrue()
 
 			attr1 := label.String("key1", "value1")
@@ -368,7 +368,7 @@ func TestSpan(t *testing.T) {
 			tracer := tp.Tracer(t.Name())
 			_, span := tracer.Start(context.Background(), "test")
 
-			subject, ok := span.(*tracetest.Span)
+			subject, ok := span.(*oteltest.Span)
 			e.Expect(ok).ToBeTrue()
 
 			expectedAttr := label.String("key", "value")
@@ -391,7 +391,7 @@ func TestSpan(t *testing.T) {
 			tracer := tp.Tracer(t.Name())
 			_, span := tracer.Start(context.Background(), "test")
 
-			subject, ok := span.(*tracetest.Span)
+			subject, ok := span.(*oteltest.Span)
 			e.Expect(ok).ToBeTrue()
 
 			var wg sync.WaitGroup
@@ -415,7 +415,7 @@ func TestSpan(t *testing.T) {
 	})
 
 	t.Run("#Links", func(t *testing.T) {
-		tp := tracetest.NewTracerProvider()
+		tp := oteltest.NewTracerProvider()
 		t.Run("returns an empty map by default", func(t *testing.T) {
 			t.Parallel()
 
@@ -424,7 +424,7 @@ func TestSpan(t *testing.T) {
 			tracer := tp.Tracer(t.Name())
 			_, span := tracer.Start(context.Background(), "test")
 
-			subject, ok := span.(*tracetest.Span)
+			subject, ok := span.(*oteltest.Span)
 			e.Expect(ok).ToBeTrue()
 
 			e.Expect(len(subject.Links())).ToEqual(0)
@@ -432,7 +432,7 @@ func TestSpan(t *testing.T) {
 	})
 
 	t.Run("#Events", func(t *testing.T) {
-		tp := tracetest.NewTracerProvider()
+		tp := oteltest.NewTracerProvider()
 		t.Run("returns an empty slice by default", func(t *testing.T) {
 			t.Parallel()
 
@@ -441,7 +441,7 @@ func TestSpan(t *testing.T) {
 			tracer := tp.Tracer(t.Name())
 			_, span := tracer.Start(context.Background(), "test")
 
-			subject, ok := span.(*tracetest.Span)
+			subject, ok := span.(*oteltest.Span)
 			e.Expect(ok).ToBeTrue()
 
 			e.Expect(len(subject.Events())).ToEqual(0)
@@ -455,7 +455,7 @@ func TestSpan(t *testing.T) {
 			tracer := tp.Tracer(t.Name())
 			_, span := tracer.Start(context.Background(), "test")
 
-			subject, ok := span.(*tracetest.Span)
+			subject, ok := span.(*oteltest.Span)
 			e.Expect(ok).ToBeTrue()
 
 			event1Name := "event1"
@@ -508,7 +508,7 @@ func TestSpan(t *testing.T) {
 			tracer := tp.Tracer(t.Name())
 			_, span := tracer.Start(context.Background(), "test")
 
-			subject, ok := span.(*tracetest.Span)
+			subject, ok := span.(*oteltest.Span)
 			e.Expect(ok).ToBeTrue()
 
 			subject.AddEvent(context.Background(), "test")
@@ -526,7 +526,7 @@ func TestSpan(t *testing.T) {
 	})
 
 	t.Run("#Status", func(t *testing.T) {
-		tp := tracetest.NewTracerProvider()
+		tp := oteltest.NewTracerProvider()
 		t.Run("defaults to OK", func(t *testing.T) {
 			t.Parallel()
 
@@ -535,7 +535,7 @@ func TestSpan(t *testing.T) {
 			tracer := tp.Tracer(t.Name())
 			_, span := tracer.Start(context.Background(), "test")
 
-			subject, ok := span.(*tracetest.Span)
+			subject, ok := span.(*oteltest.Span)
 			e.Expect(ok).ToBeTrue()
 
 			e.Expect(subject.StatusCode()).ToEqual(codes.Unset)
@@ -560,7 +560,7 @@ func TestSpan(t *testing.T) {
 				tracer := tp.Tracer(t.Name())
 				_, span := tracer.Start(context.Background(), "test")
 
-				subject, ok := span.(*tracetest.Span)
+				subject, ok := span.(*oteltest.Span)
 				e.Expect(ok).ToBeTrue()
 
 				subject.SetStatus(codes.Ok, "Ok")
@@ -578,7 +578,7 @@ func TestSpan(t *testing.T) {
 				tracer := tp.Tracer(t.Name())
 				_, span := tracer.Start(context.Background(), "test")
 
-				subject, ok := span.(*tracetest.Span)
+				subject, ok := span.(*oteltest.Span)
 				e.Expect(ok).ToBeTrue()
 
 				originalStatus := codes.Ok
@@ -594,7 +594,7 @@ func TestSpan(t *testing.T) {
 	})
 
 	t.Run("#SpanKind", func(t *testing.T) {
-		tp := tracetest.NewTracerProvider()
+		tp := oteltest.NewTracerProvider()
 		t.Run("returns the value given at start", func(t *testing.T) {
 			t.Parallel()
 
@@ -604,7 +604,7 @@ func TestSpan(t *testing.T) {
 			_, span := tracer.Start(context.Background(), "test",
 				trace.WithSpanKind(trace.SpanKindConsumer))
 
-			subject, ok := span.(*tracetest.Span)
+			subject, ok := span.(*oteltest.Span)
 			e.Expect(ok).ToBeTrue()
 			subject.End()
