+++ oteltest/tracer.go
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Package tracetest provides testing utilities for tracing.
-package tracetest
+// package oteltest provides testing utilities for tracing.
+package oteltest
 
 import (
 	"context"
